Added Component Communication
Created a Subject so that Sidebar and Article Components can communicate with each other using the common Service RedditDataService. When a Sidebar link is clicked, the Subject broadcasts that link's data to the Article Header, preventing the need for another API call.

Also, moved repetitive functions from the Sidebar-Links Component to RedditDataService.